Move MockNVPClass annd MockHandler to separate file as these are used by both PPAPIServiceTest and PPBaseServiceTest.
